ion: fix build warning
Summary: Signed-off-by: Zhang Guoying <guoying.zhang@horizon.ai>

Test Plan: no plan

Reviewers: ming.yu

Reviewed By: ming.yu

Differential Revision: https://cr.hobot.cc/D35892

+++ drivers/staging/android/ion/ion_cma_heap.c
@@ -55,7 +55,8 @@ static int ion_cma_allocate(struct ion_heap *heap, struct ion_buffer *buffer,
 		return -ENOMEM;
 
 	/* the page may used for instruction or data which still cached */
-	flush_icache_range(page_address(pages), page_address(pages) + size);
+	flush_icache_range((unsigned long)page_address(pages),
+			(unsigned long)(page_address(pages) + size));
 	__inval_dcache_area(page_address(pages), size);
 	if (PageHighMem(pages)) {
 		unsigned long nr_clear_pages = nr_pages;
